Settings + API initial
Settings:
- implemented simple settings that reads out a json file and provides as a json that can be read from and added to

APIController:
- added initial implementation to save and load API keys

Constants:
- added ui scale factors

GUI:
- changed scale change identification to use factor instead of enum
- added reading last scale from settings to always start up with the correct one
- added general, style and api tab to options window
- added font size to options window
- added api keys to options window

Logging:
- added ELogLevel to consttructors
- changed file logger to write lock the log file

Mumble:
- added GetScalingFactor based on the enum

Paths:
- added common folder
- added common/api/v2 folder
- added settings.json
- added apikeys.json

Renderer:
- changed Scaling factor to initialise as 0

+++ src/Events/EventHandler.cpp
@@ -48,7 +48,7 @@ namespace Events
 		{
 			for (auto& [identifier, consumers] : Registry)
 			{
-				int i = consumers.size() - 1;
+				size_t i = consumers.size() - 1;
 				while (i >= 0)
 				{
 					if (consumers[i] >= aStartAddress && consumers[i] <= aEndAddress)
